GETP-210 rename: config 관련 클래스를 config 컴포넌트로 분리
