+++ src/main/java/es/princip/getp/config/JpaAuditingConfig.java
@@ -1,4 +1,4 @@
-package es.princip.getp.infra.config;
+package es.princip.getp.config;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
